More package extensions (#161)
