+++ .github/workflows/CI.yml
@@ -18,7 +18,6 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.7'
           - '1'
         os:
           - ubuntu-latest
