+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ITensorNetworks"
 uuid = "2919e153-833c-4bdc-8836-1ea460a35fc7"
 authors = ["Matthew Fishman <mfishman@flatironinstitute.org> and contributors"]
-version = "0.9.0"
+version = "0.10.0"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -13,7 +13,6 @@ Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
 IsApprox = "28f27b66-4bd8-47e7-9110-e2746eb8bed7"
 IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
@@ -21,11 +20,9 @@ KrylovKit = "0b1a1467-8014-51b9-945f-bf0ae24f4b77"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NDTensors = "23ae76d9-e61a-49c4-8f12-3f1a16adf9cf"
 NamedGraphs = "678767b0-92e7-4007-89e4-4527a8725b19"
-Observers = "338f10d5-c7f1-4033-a7d1-f9dec39bcaa0"
 PackageExtensionCompat = "65ce6f38-6b18-4e1d-a461-8949797d7930"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SerializedElementArrays = "d3ce8812-9567-47e9-a7b5-65a6d70a3065"
 SimpleTraits = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
 SparseArrayKit = "a9a3c162-d163-4c15-8926-b8794fbefed2"
@@ -38,9 +35,15 @@ TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 
 [weakdeps]
 EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
+GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
+Observers = "338f10d5-c7f1-4033-a7d1-f9dec39bcaa0"
+OMEinsumContractionOrders = "6f22d1fd-8eed-4bb7-9776-e7d684900715"
 
 [extensions]
 ITensorNetworksEinExprsExt = "EinExprs"
+ITensorNetworksGraphsFlowsExt = "GraphsFlows"
+ITensorNetworksObserversExt = "Observers"
+ITensorNetworksOMEinsumContractionOrdersExt = "OMEinsumContractionOrders"
 
 [compat]
 AbstractTrees = "0.4.4"
@@ -50,32 +53,35 @@ DataGraphs = "0.2.2"
 DataStructures = "0.18"
 Dictionaries = "0.4"
 Distributions = "0.25.86"
-DocStringExtensions = "0.8, 0.9"
+DocStringExtensions = "0.9"
 EinExprs = "0.6.4"
 Graphs = "1.8"
 GraphsFlows = "0.1.1"
-ITensors = "0.3.58, 0.4"
+ITensors = "0.4"
 IsApprox = "0.1"
 IterTools = "1.4.0"
 KrylovKit = "0.6, 0.7"
 NamedGraphs = "0.5.1"
-NDTensors = "0.2, 0.3"
+NDTensors = "0.3"
 Observers = "0.2"
+OMEinsumContractionOrders = "0.8.3"
 PackageExtensionCompat = "1"
-Requires = "1.3"
 SerializedElementArrays = "0.1"
 SimpleTraits = "0.9"
-SparseArrayKit = "0.2.1, 0.3"
+SparseArrayKit = "0.3"
 SplitApplyCombine = "1.2"
 StaticArrays = "1.5.12"
 StructWalk = "0.2"
 Suppressor = "0.2"
 TimerOutputs = "0.5.22"
 TupleTools = "1.4"
-julia = "1.7"
+julia = "1.10"
 
 [extras]
 EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
+GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
+Observers = "338f10d5-c7f1-4033-a7d1-f9dec39bcaa0"
+OMEinsumContractionOrders = "6f22d1fd-8eed-4bb7-9776-e7d684900715"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
