+++ src/apply.jl
@@ -27,7 +27,6 @@ using ITensors.ITensorMPS: siteinds
 using KrylovKit: linsolve
 using LinearAlgebra: eigen, norm, svd
 using NamedGraphs: NamedEdge, has_edge
-using Observers: Observers
 
 function full_update_bp(
   o,
