--- src/observers.jl
@@ -1,7 +1,0 @@
-# TODO: Move to `ITensorNetworksObserversExt`.
-using Observers: Observers
-
-"""
-Overload of `Observers.update!`.
-"""
-Observers.update!(::Nothing; kwargs...) = nothing
