+++ test/Project.toml
@@ -12,7 +12,6 @@ ITensorGaussianMPS = "2be41995-7c9f-4653-b682-bfa4e7cebb93"
 ITensorNetworks = "2919e153-833c-4bdc-8836-1ea460a35fc7"
 ITensorUnicodePlots = "73163f41-4a9e-479f-8353-73bf94dbd758"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
-KaHyPar = "2a6221f6-aa48-11e9-3542-2d9e0ef01880"
 KrylovKit = "0b1a1467-8014-51b9-945f-bf0ae24f4b77"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
